@@ -1,7 +1,7 @@
-require 'spec_helper'
+require "spec_helper"
 
 describe FastJsonapi::ObjectSerializer do
-  include_context 'movie class'
+  include_context "movie class"
 
   context "params option" do
     let(:hash) { serializer.serializable_hash }
@@ -18,7 +18,7 @@ describe FastJsonapi::ObjectSerializer do
           params[:user] ? movie.viewed?(params[:user]) : false
         end
 
-        attribute :no_param_attribute do |movie|
+        attribute :no_param_attribute do |_movie|
           "no-param-attribute"
         end
       end
@@ -34,34 +34,34 @@ describe FastJsonapi::ObjectSerializer do
       let(:params) { User.new([]) }
 
       it "fails when creating a serializer with an object as params" do
-        expect(-> { MovieSerializer.new(movie, {params: User.new([])}) }).to raise_error(ArgumentError)
+        expect(-> { MovieSerializer.new(movie, params: User.new([])) }).to raise_error(ArgumentError)
       end
 
       it "succeeds creating a serializer with a hash" do
-        expect(-> { MovieSerializer.new(movie, {params: {current_user: User.new([])}}) }).not_to raise_error
+        expect(-> { MovieSerializer.new(movie, params: { current_user: User.new([]) }) }).not_to raise_error
       end
     end
 
     context "passing params to the serializer" do
-      let(:params) { {user: User.new([movie.id])} }
-      let(:options_with_params) { {params: params} }
+      let(:params) { { user: User.new([movie.id]) } }
+      let(:options_with_params) { { params: params } }
 
       context "with a single record" do
         let(:serializer) { MovieSerializer.new(movie, options_with_params) }
 
         it "handles attributes that use params" do
-          expect(hash[:data][:attributes][:viewed]).to eq(true)
+          expect(hash[:data][:viewed]).to eq(true)
         end
 
         it "handles attributes that don't use params" do
-          expect(hash[:data][:attributes][:no_param_attribute]).to eq("no-param-attribute")
+          expect(hash[:data][:no_param_attribute]).to eq("no-param-attribute")
         end
       end
 
       context "with a list of records" do
         let(:movies) { build_movies(3) }
         let(:user) { User.new(movies.map { |m| [true, false].sample ? m.id : nil }.compact) }
-        let(:params) { {user: user} }
+        let(:params) { { user: user } }
         let(:serializer) { MovieSerializer.new(movies, options_with_params) }
 
         it "has 3 items" do
@@ -69,14 +69,14 @@ describe FastJsonapi::ObjectSerializer do
         end
 
         it "handles passing params to a list of resources" do
-          param_attribute_values = hash[:data].map { |data| [data[:id], data[:attributes][:viewed]] }
-          expected_values = movies.map { |m| [m.id.to_s, user.viewed.include?(m.id)] }
+          param_attribute_values = hash[:data].map { |data| [data[:id], data[:viewed]] }
+          expected_values = movies.map { |m| [m.id, user.viewed.include?(m.id)] }
 
           expect(param_attribute_values).to eq(expected_values)
         end
 
         it "handles attributes without params" do
-          no_param_attribute_values = hash[:data].map { |data| data[:attributes][:no_param_attribute] }
+          no_param_attribute_values = hash[:data].map { |data| data[:no_param_attribute] }
           expected_values = (1..3).map { "no-param-attribute" }
 
           expect(no_param_attribute_values).to eq(expected_values)
@@ -89,11 +89,11 @@ describe FastJsonapi::ObjectSerializer do
         let(:serializer) { MovieSerializer.new(movie) }
 
         it "handles param attributes" do
-          expect(hash[:data][:attributes][:viewed]).to eq(false)
+          expect(hash[:data][:viewed]).to eq(false)
         end
 
         it "handles attributes that don't use params" do
-          expect(hash[:data][:attributes][:no_param_attribute]).to eq("no-param-attribute")
+          expect(hash[:data][:no_param_attribute]).to eq("no-param-attribute")
         end
       end
 
@@ -101,13 +101,13 @@ describe FastJsonapi::ObjectSerializer do
         let(:serializer) { MovieSerializer.new(build_movies(3)) }
 
         it "handles attributes with params" do
-          param_attribute_values = hash[:data].map { |data| data[:attributes][:viewed] }
+          param_attribute_values = hash[:data].map { |data| data[:viewed] }
 
           expect(param_attribute_values).to eq([false, false, false])
         end
 
         it "handles attributes that don't use params" do
-          no_param_attribute_values = hash[:data].map { |data| data[:attributes][:no_param_attribute] }
+          no_param_attribute_values = hash[:data].map { |data| data[:no_param_attribute] }
           expected_attribute_values = (1..3).map { "no-param-attribute" }
 
           expect(no_param_attribute_values).to eq(expected_attribute_values)
