@@ -1,7 +1,6 @@
-require 'spec_helper'
+require "spec_helper"
 
 describe FastJsonapi::ObjectSerializer do
-
   after(:all) do
     classes_to_remove = %i[
       User
@@ -41,7 +40,7 @@ describe FastJsonapi::ObjectSerializer do
     set_type :user
     attributes :first_name, :last_name
 
-    attribute :full_name do |user, params|
+    attribute :full_name do |user, _params|
       "#{user.first_name} #{user.last_name}"
     end
 
@@ -95,74 +94,71 @@ describe FastJsonapi::ObjectSerializer do
     has_one :account
   end
 
-  it 'sets the correct record type' do
+  it "sets the correct record type" do
     expect(EmployeeSerializer.reflected_record_type).to eq :employee
     expect(EmployeeSerializer.record_type).to eq :employee
   end
 
-  context 'when testing inheritance of attributes' do
-
-    it 'includes parent attributes' do
+  context "when testing inheritance of attributes" do
+    it "includes parent attributes" do
       subclass_attributes = EmployeeSerializer.attributes_to_serialize
       superclass_attributes = UserSerializer.attributes_to_serialize
       expect(subclass_attributes).to include(superclass_attributes)
     end
 
-    it 'returns inherited attribute with a block correctly' do
+    it "returns inherited attribute with a block correctly" do
       e = Employee.new
       e.id = 1
-      e.first_name = 'S'
-      e.last_name = 'K'
-      attributes_hash = EmployeeSerializer.new(e).serializable_hash[:data][:attributes]
-      expect(attributes_hash).to include(full_name: 'S K')
+      e.first_name = "S"
+      e.last_name = "K"
+      attributes_hash = EmployeeSerializer.new(e).serializable_hash[:data]
+      expect(attributes_hash).to include(full_name: "S K")
     end
 
-    it 'includes child attributes' do
+    it "includes child attributes" do
       expect(EmployeeSerializer.attributes_to_serialize[:location].method).to eq(:location)
     end
 
-    it 'doesnt change parent class attributes' do
+    it "doesnt change parent class attributes" do
       EmployeeSerializer
       expect(UserSerializer.attributes_to_serialize).not_to have_key(:location)
     end
   end
 
-  context 'when testing inheritance of relationship' do
-    it 'includes parent relationships' do
+  context "when testing inheritance of relationship" do
+    it "includes parent relationships" do
       subclass_relationships = EmployeeSerializer.relationships_to_serialize
       superclass_relationships = UserSerializer.relationships_to_serialize
       expect(subclass_relationships).to include(superclass_relationships)
     end
 
-    it 'returns inherited relationship correctly' do
+    it "returns inherited relationship correctly" do
       e = Employee.new
       e.country_id = 1
       relationships_hash = EmployeeSerializer.new(e).serializable_hash[:data][:relationships][:country]
-      expect(relationships_hash).to include(data: { id: "1", type: :country })
+      expect(relationships_hash).to include(data: { id: 1 })
     end
 
-    it 'includes child relationships' do
+    it "includes child relationships" do
       expect(EmployeeSerializer.relationships_to_serialize.keys).to include(:account)
     end
 
-    it 'doesnt change parent class attributes' do
+    it "doesnt change parent class attributes" do
       EmployeeSerializer
       expect(UserSerializer.relationships_to_serialize.keys).not_to include(:account)
     end
 
-    it 'includes parent cached relationships' do
+    it "includes parent cached relationships" do
       subclass_relationships = EmployeeSerializer.cachable_relationships_to_serialize
       superclass_relationships = UserSerializer.cachable_relationships_to_serialize
       expect(subclass_relationships).to include(superclass_relationships)
     end
   end
 
-  context 'when test inheritence of other attributes' do
-
-    it 'inherits the tranform method' do
+  context "when test inheritence of other attributes" do
+    it "inherits the tranform method" do
       EmployeeSerializer
       expect(UserSerializer.transform_method).to eq EmployeeSerializer.transform_method
     end
-
   end
 end
